pyproject.toml
[project]
name = "valorant-premier-analyst"
version = "0.1.0"
description = "Valorant Premier team match history analysis and lightweight AI analyst toolkit."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Valorant Premier Analyst" }]
license = { text = "MIT" }

dependencies = [
    "pandas>=2.2.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.1",
    "duckdb>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "types-requests>=2.32.0",
    "httpx>=0.27.0",
]

[project.scripts]
valorant-analyst = "valorant_analyst.cli:main"
valorant-analyst-server = "valorant_analyst.server.app:run"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/valorant_analyst"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B"]
# FastAPI idiom: `Depends(...)`, `Query(...)`, etc. as default values is intentional.
"src/valorant_analyst/server/routes/*" = ["B008"]
"src/valorant_analyst/server/deps.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
files = ["src"]
